moved files to their correct subdir, restored root makefile
