+++ makefile
@@ -1,59 +1,17 @@
 CC = gcc
 
-CFLAGS = -Wall -Werror -Wextra -ggdb3
-# I_FLAGS = $(shell pkg-config --cflags gtk4)
-# L_FLAGS = $(shell pkg-config --libs gtk4)
+CFLAGS = -Wall -Werror -Wextra -pedantic -ggdb3
+# I_FLAGS += $(shell pkg-config --cflags gtk4) # include flags
+# L_FLAGS += $(shell pkg-config --libs gtk4)   # link flags
 
 CCMD = ${CC} ${I_FLAGS} ${CFLAGS} ${L_FLAGS}
 
 .DELETE_ON_ERROR:
 
-.PHONY: clean
+.PHONY: clean 
 
-
-all: pid ppid noargc readline splitter execve forker wait five simpsh which
-
-pid:
-	${CCMD} pid.c -o pid.exec
-
-ppid:
-	${CCMD} ppid.c -o ppid.exec
-
-noargc:
-	${CCMD} noargc.c -o noargc.exec
-
-readline:
-	${CCMD} readline.c -o readline.exec
-
-splitter:
-	${CCMD} splitter.c -o splitter.exec
-
-execve:
-	${CCMD} execve.c -o execve.exec
-
-forker:
-	${CCMD} forker.c -o forker.exec
-
-wait:
-	${CCMD} wait.c -o wait.exec
-
-five:
-	${CCMD} exec5.c -o exec5.exec
-
-stat:
-	${CCMD} stat.c -o stat.exec
-
-envmain:
-	${CCMD} envmain.c -o envmain.exec
-
-environ:
-	${CCMD} environ.c -o environ.exec
-
-which:
-	${CCMD} _which.c -o _which.exec
-
-simpsh:
-	${CCMD} simpsh.c -o simpsh.exec
+all: 
+	${CCMD} *.c -o hsh 
 
 clean:
-	$(RM) *~ *.exec *.o
+	$(RM) *~ *.exec *.o hsh
